TeleKB/requirements.txt
telethon
google-genai
python-dotenv
tk
